--- nextjs-fundamental/src/app/achmadshobari/components.tsx/profile.tsx
@@ -1,7 +1,0 @@
-export default function Profile() {
-  return (
-    <div className="relative r-0">
-      <h1>test</h1>
-    </div>
-  );
-}
