--- nextjs-fundamental/src/app/achmadshobari/components.tsx/sidebar.tsx
@@ -1,159 +1,0 @@
-export default function SidebarComponent() {
-  return (
-    <>
-      <header className="absolute top-0 left-0 z-10 flex w-full items-center bg-transparent">
-        <div className="container">
-          <div className="relative flex items-center justify-between">
-            <div className="px-4">
-              <a
-                href="#home"
-                className="block py-6 text-lg font-bold text-primary"
-              >
-                Achmad
-              </a>
-            </div>
-            <div className="flex items-center px-4">
-              <button
-                id="hamburger"
-                name="hamburger"
-                type="button"
-                className="absolute right-4 block lg:hidden"
-              >
-                <span className="hamburger-line origin-top-left transition duration-300 ease-in-out"></span>
-                <span className="hamburger-line transition duration-300 ease-in-out"></span>
-                <span className="hamburger-line origin-bottom-left transition duration-300 ease-in-out"></span>
-              </button>
-
-              <nav
-                id="nav-menu"
-                className="absolute right-4 top-full hidden w-full max-w-[250px] rounded-lg bg-white py-5 shadow-lg dark:bg-dark dark:shadow-slate-500 lg:static lg:block lg:max-w-full lg:rounded-none lg:bg-transparent lg:shadow-none lg:dark:bg-transparent"
-              >
-                <ul className="block lg:flex">
-                  <li className="group">
-                    <a
-                      href="#home"
-                      className="mx-8 flex py-2 text-base text-dark group-hover:text-primary dark:text-white"
-                    >
-                      Beranda
-                    </a>
-                  </li>
-                  <li className="group">
-                    <a
-                      href="#about"
-                      className="mx-8 flex py-2 text-base text-dark group-hover:text-primary dark:text-white"
-                    >
-                      Tentang Saya
-                    </a>
-                  </li>
-                  <li className="group">
-                    <a
-                      href="#portfolio"
-                      className="mx-8 flex py-2 text-base text-dark group-hover:text-primary dark:text-white"
-                    >
-                      Portfolio
-                    </a>
-                  </li>
-                  <li className="group">
-                    <a
-                      href="#clients"
-                      className="mx-8 flex py-2 text-base text-dark group-hover:text-primary dark:text-white"
-                    >
-                      Clients
-                    </a>
-                  </li>
-                  <li className="group">
-                    <a
-                      href="#blog"
-                      className="mx-8 flex py-2 text-base text-dark group-hover:text-primary dark:text-white"
-                    >
-                      Blog
-                    </a>
-                  </li>
-                  <li className="group">
-                    <a
-                      href="#contact"
-                      className="mx-8 flex py-2 text-base text-dark group-hover:text-primary dark:text-white"
-                    >
-                      Contact
-                    </a>
-                  </li>
-                  <li className="mt-3 flex items-center pl-8 lg:mt-0">
-                    <div className="flex">
-                      <span className="mr-2 text-sm text-slate-500">light</span>
-                      <input
-                        type="checkbox"
-                        className="hidden"
-                        id="dark-toggle"
-                      />
-                      <label htmlFor="dark-toggle">
-                        <div className="flex h-5 w-9 cursor-pointer items-center rounded-full bg-slate-500 p-1">
-                          <div className="toggle-circle h-4 w-4 rounded-full bg-white transition duration-300 ease-in-out"></div>
-                        </div>
-                      </label>
-                      <span className="ml-2 text-sm text-slate-500">dark</span>
-                    </div>
-                  </li>
-                </ul>
-              </nav>
-            </div>
-          </div>
-        </div>
-      </header>
-
-      <section id="home" className="pt-36 dark:bg-dark">
-        <div className="container">
-          <div className="flex flex-wrap">
-            <div className="w-full self-center px-4 lg:w-1/2">
-              <h1 className="text-base font-semibold text-primary md:text-xl">
-                Halo Semua 👋, saya{" "}
-                <span className="mt-1 block text-4xl font-bold text-dark dark:text-white lg:text-5xl">
-                  Achmad
-                </span>
-              </h1>
-              <h2 className="mb-5 text-lg font-medium text-secondary lg:text-2xl">
-                Student &{" "}
-                <span className="text-dark dark:text-white">Web developer</span>
-              </h2>
-              <p className="mb-10 font-medium leading-relaxed text-secondary">
-                Belajar web programming itu mudah dan menyenangkan bukan.{" "}
-                <span className="font-bold text-dark dark:text-white">
-                  bukan!
-                </span>
-              </p>
-
-              <a
-                href="#"
-                className="rounded-full bg-primary py-3 px-8 text-base font-semibold text-white transition duration-300 ease-in-out hover:opacity-80 hover:shadow-lg"
-              >
-                Hubungi Saya
-              </a>
-            </div>
-            <div className="w-full self-end px-4 lg:w-1/2">
-              <div className="relative mt-10 lg:right-0 lg:mt-9">
-                <img
-                  src=""
-                  alt=""
-                  className="relative z-10 mx-auto max-w-full"
-                />
-                <span className="absolute bottom-0 left-1/2 -translate-x-1/2 md:scale-125">
-                  <svg
-                    width="400"
-                    height="400"
-                    viewBox="0 0 200 200"
-                    xmlns="http://www.w3.org/2000/svg"
-                  >
-                    <path
-                      fill="#14b8a6"
-                      d="M47.4,-51C59.2,-35.6,65.1,-17.8,63.4,-1.7C61.7,14.5,52.6,28.9,40.8,36.6C28.9,44.2,14.5,45,-2,47C-18.4,49,-36.8,52.1,-45.4,44.5C-53.9,36.8,-52.5,18.4,-51.2,1.3C-49.9,-15.8,-48.7,-31.7,-40.2,-47.1C-31.7,-62.4,-15.8,-77.3,1,-78.3C17.8,-79.2,35.6,-66.3,47.4,-51Z"
-                      transform="translate(100 100) scale(1.1)"
-                    />
-                  </svg>
-                </span>
-              </div>
-            </div>
-          </div>
-        </div>
-      </section>
-    </>
-  );
-}
